refactor(lib/rule): move rules logic to rules mod
