+++ lib/src/lib.rs
@@ -1,4 +1,6 @@
 #[cfg(feature = "transaction")]
 pub mod date;
+#[cfg(feature = "rule")]
+pub mod rules;
 pub mod statements;
 pub mod value;
